change from Int64 to Int

+++ src/types.jl
@@ -21,7 +21,7 @@ Parameters for the rotating detonation engine (RDE) model.
 - `x0::T`: Initial position
 """
 @kwdef mutable struct RDEParam{T <: AbstractFloat}
-    N::Int64 = 512               # Number of spatial points
+    N::Int = 512               # Number of spatial points
     L::T = 2.0f0π                  # Domain length
     ν_1::T = 0.0075f0           # Viscosity coefficient
     ν_2::T = 0.0075f0
@@ -29,7 +29,7 @@ Parameters for the rotating detonation engine (RDE) model.
     α::T = 0.3f0                # Parameter in ω(u)
     q_0::T = 1.0f0              # Source term coefficient
     u_0::T = 0.0f0              # Parameter in ξ(u, u_0)
-    n::Int64 = 1                  # Exponent in ξ(u, u_0)
+    n::Int = 1                  # Exponent in ξ(u, u_0)
     k_param::T = 5.0f0          # Parameter in β(u, s)
     u_p::T = 0.5f0              # Parameter in β(u, s)
     s::T = 3.5f0                # Parameter in β(u, s)
@@ -177,7 +177,7 @@ mutable struct FDRDECache{T <: AbstractFloat} <: AbstractRDECache{T}
     ξu::Vector{T}
     βu::Vector{T}
     dx::T
-    N::Int64
+    N::Int
     u_p_current::Vector{T}
     u_p_previous::Vector{T}
     τ_smooth::T
@@ -314,7 +314,7 @@ mutable struct UpwindRDECache{T <: AbstractFloat} <: AbstractRDECache{T}
     ξu::Vector{T}
     βu::Vector{T}
     dx::T
-    N::Int64
+    N::Int
     u_p_current::Vector{T}
     u_p_previous::Vector{T}
     τ_smooth::T
@@ -405,7 +405,7 @@ mutable struct FVCache{T <: AbstractFloat} <: AbstractRDECache{T}
     F̂::Vector{T}
     adv::Vector{T}
     dx::T
-    N::Int64
+    N::Int
     u_p_current::Vector{T}
     u_p_previous::Vector{T}
     τ_smooth::T
